+++ app/src/ai/document/mod.rs
@@ -1,2 +1,3 @@
 pub mod ai_document_model;
 pub mod orchestration_config_block;
+pub(in crate::ai) mod plan_publication;
